app/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
boto3==1.22.13
botocore==1.25.13
celery==5.2.7
certifi==2021.10.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.2.0
Deprecated==1.2.13
Django==4.0.4
django-appconf==1.0.5
django-celery-beat==2.3.0
django-celery-email==3.0.0
django-ckeditor==6.4.0
django-environ==0.9.0
django-js-asset==2.0.0
django-storages==1.12.3
django-timezone-field==5.0
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==13.14.0
flake8==4.0.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
jmespath==1.0.0
kombu==5.2.4
mccabe==0.6.1
mock==4.0.3
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2022.1
redis==4.3.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0

